refactor(podman): extract `WinInstaller` to a dedicated file
