+++ extensions/podman/packages/extension/src/utils/podman-install.spec.ts
@@ -28,7 +28,7 @@ import { releaseNotes } from '../podman5.json';
 import { getBundledPodmanVersion } from './podman-bundled';
 import type { InstalledPodman } from './podman-cli';
 import type { PodmanInfo, UpdateCheck } from './podman-install';
-import { PodmanInstall, WinInstaller } from './podman-install';
+import { PodmanInstall } from './podman-install';
 import * as utils from './util';
 
 const originalConsoleError = console.error;
@@ -67,6 +67,9 @@ const provider: extensionApi.Provider = {
   updateWarnings: vi.fn(),
   onDidUpdateDetectionChecks: vi.fn(),
 };
+
+// mock filesystem
+vi.mock('node:fs');
 
 // mock ps-list
 vi.mock('ps-list', async () => {
@@ -131,11 +134,6 @@ vi.mock(import('./util'), async () => {
   };
 });
 
-const progress = {
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  report: (): void => {},
-};
-
 beforeEach(() => {
   vi.clearAllMocks();
   // reset array of subscriptions
@@ -150,62 +148,6 @@ beforeEach(() => {
 
 afterEach(() => {
   console.error = originalConsoleError;
-});
-
-test('expect update on windows to show notification in case of 0 exit code', async () => {
-  vi.spyOn(extensionApi.window, 'withProgress').mockImplementation((options, task) => {
-    return task(progress, {} as unknown as extensionApi.CancellationToken);
-  });
-
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => Promise.resolve({} as extensionApi.RunResult));
-
-  vi.mock('node:fs');
-  vi.spyOn(fs, 'existsSync').mockReturnValue(true);
-  vi.mocked(fs.readdirSync).mockReturnValue([]);
-
-  const installer = new WinInstaller(extensionContext);
-  const result = await installer.update();
-  expect(result).toBeTruthy();
-  expect(extensionApi.window.showNotification).toHaveBeenCalled();
-});
-
-test('expect update on windows not to show notification in case of 1602 exit code', async () => {
-  vi.spyOn(extensionApi.window, 'withProgress').mockImplementation((options, task) => {
-    return task(progress, {} as unknown as extensionApi.CancellationToken);
-  });
-  const customError = { exitCode: 1602 } as extensionApi.RunError;
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-    throw customError;
-  });
-
-  vi.mock('node:fs');
-  vi.spyOn(fs, 'existsSync').mockReturnValue(true);
-  vi.mocked(fs.readdirSync).mockReturnValue([]);
-
-  const installer = new WinInstaller(extensionContext);
-  const result = await installer.update();
-  expect(result).toBeTruthy();
-  expect(extensionApi.window.showNotification).not.toHaveBeenCalled();
-});
-
-test('expect update on windows to throw error if non zero exit code', async () => {
-  vi.spyOn(extensionApi.window, 'withProgress').mockImplementation((options, task) => {
-    return task(progress, {} as unknown as extensionApi.CancellationToken);
-  });
-  const customError = { exitCode: -1, stderr: 'CustomError' } as extensionApi.RunError;
-
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-    throw customError;
-  });
-
-  vi.mock('node:fs');
-  vi.spyOn(fs, 'existsSync').mockReturnValue(true);
-  vi.mocked(fs.readdirSync).mockReturnValue([]);
-
-  const installer = new WinInstaller(extensionContext);
-  const result = await installer.update();
-  expect(result).toBeFalsy();
-  expect(extensionApi.window.showErrorMessage).toHaveBeenCalled();
 });
 
 class TestPodmanInstall extends PodmanInstall {
@@ -517,6 +459,17 @@ describe('performUpdate', () => {
 });
 
 describe('MacOSInstaller', () => {
+  // call the parameter
+  const progress = {
+    report: vi.fn(),
+  };
+
+  beforeEach(() => {
+    vi.mocked(extensionApi.window.withProgress).mockImplementation((options, task) => {
+      return task(progress, {} as unknown as extensionApi.CancellationToken);
+    });
+  });
+
   test('call installer if universal installer is found', async () => {
     // say we're using macOS
     vi.spyOn(os, 'platform').mockReturnValue('darwin');
@@ -545,11 +498,6 @@ describe('MacOSInstaller', () => {
     const methodArgs = withProgressMock.mock.calls[0];
     const promiseArg = methodArgs[1];
     expect(promiseArg).toBeDefined();
-
-    // call the parameter
-    const progress = {
-      report: vi.fn(),
-    };
 
     const token = {
       isCancellationRequested: false,
